chore: Move libp2p settings into a dedicated module.
